feat: rename file to Language.ts
