+++ packages/app-bridge/src/types/index.ts
@@ -21,11 +21,11 @@ export * from './FileExtension';
 export * from './FileExtensionSets';
 export * from './FileType';
 export * from './GuidelineSearchResult';
+export * from './Language';
 export * from './Notify';
 export * from './OauthTokens';
 export * from './PageTemplateAsset';
 export * from './PrivacySettings';
-export * from './Portal';
 export * from './PostExternalAssetParams';
 export * from './Project';
 export * from './SubscriptionHandler';
